% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawCosts.R
\name{calculateRawAvgCosts}
\alias{calculateRawAvgCosts}
\title{Calculate the raw average annual cost of invasions over periods of time}
\usage{
calculateRawAvgCosts(
  costdb,
  cost.column = "Cost_estimate_per_year_2017_USD_exchange_rate",
  year.column = "Impact_year",
  in.millions = TRUE,
  cost.transf = "log10",
  minimum.year = 1960,
  maximum.year = 2017,
  year.breaks = seq(minimum.year, maximum.year, by = 10),
  plotbreaks = c(0.1, 1, 10, 100, 1000, 10000, 1e+05, 1e+06, 1e+07, 1e+08, 1e+09, 1e+10,
    1e+11, 1e+12)
)
}
\arguments{
\item{costdb}{The \bold{expanded INVACOST database} output from 
\code{\link{expandYearlyCosts}},
where annual costs occurring over several years are repeated for each year.}

\item{cost.column}{Name of the cost column to use in \code{costdb} (usually, 
choose between the exchange rate (default) or PPP annualised cost)}

\item{year.column}{Name of the year column to use in \code{costdb}.}

\item{in.millions}{If \code{TRUE}, cost values will be transformed in 
millions (to make graphs easier to read), else if \code{}, cost values will
not be transformed.}

\item{cost.transf}{Type of transformation you want to apply on cost values.
Specify \code{NA} or \code{NULL} to avoid any transformation. Only useful
for graphical representation.}

\item{minimum.year}{the starting year of this analysis. By default, 
1960 was chosen because it marks the period from which world bank data is 
available for exchange rates and inflation values.}

\item{maximum.year}{the ending year for this analysis. By default, 2017
was chosen as it is the last year for which we have data in INVACOST.}

\item{year.breaks}{a vector of breaks for the year intervals over which
you want to calculate raw cost values}

\item{plot.breaks}{a vector of numeric values indicating the plot breaks 
for the axis of average annual cost values.}
}
\value{
a \code{list} with 5 elements:

\itemize{
\item{\code{cost.data}: the annualised costs of invasions, as sums of all 
costs for each year.}
\item{\code{parameters}: parameters used to run the function. The 
\code{minimum.year} and \code{maximum.year} are based on the input data
(i.e., the user may specify \code{minimum.year = 1960} but the input data may
only have data starting from 1970, hence the \code{minimum.year} will be
 1970.)}
\item{\code{average.total.cost}: the average annual cost of IAS calculated
over the entire time period}
\item{\code{average.cost.per.period}: a data.frame containing the the average 
annual cost of IAS calculated over each time interval}
\item{\code{plot}}: the ggplot object of the output plot.
}
The structure of this object can be seen using \code{str()}
}
\description{
This function calculates the raw average annual cost of invasive species
over different periods of time
}
\details{

}
\examples{
# Create an example stack with two environmental variables
data(invacost)
db.over.time <- expandYearlyCosts(invacost,
                                  startcolumn = "Probable_Starting_year_Low_margin",
                                  endcolumn = "Probable_Ending_year_Low_margin")
costdb <- db.over.time[db.over.time$Implementation == "Observed", ]
costdb <- costdb[which(costdb$Method_reliability == "High"), ]
res <- calculateRawAvgCosts(costdb)
res
}
\seealso{
\code{\link{expandYearlyCosts}} to get the database in appropriate format.
}
\author{
Boris Leroy \email{leroy.boris@gmail.com}

with help from C. Diagne & A.-C. Vaissière
}
