% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAnnualCostValue.R
\name{estimateAnnualCosts}
\alias{estimateAnnualCosts}
\title{Estimate the average annual cost value of invasions}
\usage{
estimateAnnualCosts(
  costdb,
  cost.column = "Cost_estimate_per_year_2017_USD_exchange_rate",
  year.column = "Impact_year",
  cost.transf = "log10",
  in.millions = TRUE,
  confidence.interval = 0.95,
  plotbreaks = c(0.1, 1, 10, 100, 1000, 10000, 1e+05, 1e+06, 1e+07, 1e+08, 1e+09, 1e+10,
    1e+11, 1e+12),
  minimum.year = 1960,
  maximum.year = 2017,
  final.year = 2017,
  incomplete.year.threshold = 2010,
  incomplete.year.weights = NULL
)
}
\arguments{
\item{costdb}{The \bold{expanded INVACOST database} output from 
\code{\link{expandYearlyCosts}},
where annual costs occurring over several years are repeated for each year.}

\item{cost.column}{Name of the cost column to use in \code{costdb} (usually, 
choose between the exchange rate (default) or PPP annualised cost)}

\item{year.column}{Name of the year column to use in \code{costdb}.}

\item{cost.transf}{Type of transformation you want to apply on cost values.
The default is a log10 transformation, which is commonly applied in economy,
allows to fit linear regression with a normal distribution of residuals, 
and makes plots easy to read. You can apply another transformation by 
specifying the name of the transformation function (e.g., natural
logarithm, {\code{"log"}}). Specify \code{NA} or \code{NULL} to avoid any 
transformation.}

\item{in.millions}{If \code{TRUE}, cost values will be transformed in 
millions (to make graphs easier to read), else if \code{}, cost values will
not be transformed.}

\item{minimum.year}{the starting year of this analysis. By default, 
1960 was chosen because it marks the period from which world bank data is 
available for exchange rates and inflation values.}

\item{maximum.year}{the ending year for this analysis. By default, 2017
was chosen as it is the last year for which we have data in INVACOST.}

\item{incomplete.year.threshold}{Estimated threshold for incomplete cost 
data. All years above or equal to this threshold will be excluded from 
model calibration, because of the time-lag between economic impacts of
invasive species and the documentation and publication of these impacts.}

\item{incomplete.year.weights}{A named vector containing weights of years
for the regressions. Useful to decrease the weights of incomplete years
in regressions. Names of this vector must correspond to years.}

\item{confidence.intervals}{a numeric value between 0 and 1, corresponding
to the desired confidence intervals around model predictions.}

\item{plot.breaks}{a vector of numeric values indicating the plot breaks 
for the axis of annual cost values.}

\item{final.year.cost}{the year for which you want to obtain the final 
average cost estimate from models. Default is 2017.}
}
\value{
a \code{list} with 3 to 6 elements (only the first three will be 
provided if you selected a cost transformation different from log10):

\itemize{
\item{\code{cost.data}: the annualised costs of invasions, as sums of all 
costs for each year.}
\item{\code{parameters}: parameters used to run the function. The 
\code{minimum.year} and \code{maximum.year} are based on the input data
(i.e., the user may specify \code{minimum.year = 1960} but the input data may
only have data starting from 1970, hence the \code{minimum.year} will be
 1970.)}
\item{\code{fitted.models}: a list of objects the fitted models.}
\item{\code{estimated.annual.costs}}: a data.frame containing the predicted 
cost values for each year for all the fitted models.
\item{\code{RMSE}}: an array containing RMSE of models for the calibration 
data and for all data. NOTE: the RMSE for quantile regression is not an 
relevant metric. 
\item{\code{plot}}: the ggplot object of the output plot.
\item{\code{final.year.cost}}: a vector containing the estimated annual
costs of invasive species based on all models for \code{final.year}.
}
The structure of this object can be seen using \code{str()}
}
\description{
This function fits different models on annualised INVACOST data in order to
estimate the average annual cost of invasive species.
}
\details{

}
\examples{
# Create an example stack with two environmental variables
data(invacost)
db.over.time <- expandYearlyCosts(invacost,
                                  startcolumn = "Probable_Starting_year_Low_margin",
                                  endcolumn = "Probable_Ending_year_Low_margin")
costdb <- db.over.time[db.over.time$Implementation == "Observed", ]
costdb <- costdb[which(costdb$Method_reliability == "High"), ]
costdb <- costdb[-which(costdb$Reference_ID == 8733), ]
res <- estimateAnnualCosts(costdb)
}
\seealso{
\code{\link{expandYearlyCosts}} to get the database in appropriate format.
}
\author{
Boris Leroy \email{leroy.boris@gmail.com}

with help from C. Diagne & A.-C. Vaissière
}
