% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRawCosts.R, R/costTrendOverTime.R,
%   R/invacost-deprecated.R
\name{invacost-deprecated}
\alias{invacost-deprecated}
\alias{calculateRawAvgCosts}
\alias{rawAvgCost}
\alias{costTrendOverTime}
\title{Deprecated functions in package \pkg{invacost}.}
\usage{
calculateRawAvgCosts(
  costdb,
  cost.column = "Cost_estimate_per_year_2017_USD_exchange_rate",
  year.column = "Impact_year",
  in.millions = TRUE,
  minimum.year = 1960,
  maximum.year = 2017,
  year.breaks = seq(minimum.year, maximum.year, by = 10),
  include.last.year = TRUE
)

rawAvgCost(
  costdb,
  cost.column = "Cost_estimate_per_year_2017_USD_exchange_rate",
  year.column = "Impact_year",
  min.year = NULL,
  max.year = NULL
)

costTrendOverTime(
  costdb,
  cost.column = "Cost_estimate_per_year_2017_USD_exchange_rate",
  year.column = "Impact_year",
  cost.transf = "log10",
  in.millions = TRUE,
  confidence.interval = 0.95,
  minimum.year = 1960,
  maximum.year = 2017,
  final.year = 2017,
  incomplete.year.threshold = NULL,
  incomplete.year.weights = NULL,
  gam.k = -1,
  mars.nprune = NULL,
  ...
)
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. Alternative functions with similar
  functionality are provided. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\section{\code{calculateRawAvgCosts}}{

For \code{calculateRawAvgCosts}, use \code{\link{summarizeCosts}}.
}

\section{\code{rawAvgCost}}{

For \code{rawAvgCost}, use \code{\link{computeAvgTotCost}}.
}

\section{\code{costTrendOverTime}}{

For \code{costTrendOverTime}, use \code{\link{modelCosts}}.
}

\keyword{internal}
