% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invacost.R
\docType{data}
\name{invacost}
\alias{invacost}
\title{INVACOST database}
\format{A data frame with 2419 rows and 48 variables:
\describe{
  \item{Cost_ID}{}
  \item{Repository}{}
  \item{Reference_ID}{}
  \item{Reference_title}{}
  \item{Authors}{}
  \item{Publication_year}{}
  \item{Type_of_material}{}
  \item{Previous_materials}{}
  \item{Availability}{}
  \item{Kingdom}{}
  \item{Phylum}{}
  \item{Class}{}
  \item{Order}{}
  \item{Family}{}
  \item{Genus}{}
  \item{Species}{}
  \item{Common_name}{}
  \item{Environment}{}
  \item{Geographic_region}{}
  \item{Official_country}{}
  \item{Location}{}
  \item{Spatial_scale}{}
  \item{Period_of_estimation}{}
  \item{Time_range}{}
  \item{Probable_starting_year}{}
  \item{Probable_ending_year}{}
  \item{Occurrence}{}
  \item{Raw_cost_estimate_local_currency}{}
  \item{Min_Raw_cost_estimate_local_currency}{}
  \item{Max_Raw_cost_estimate_local_currency}{}
  \item{Raw_cost_estimate_2017_USD_exchange_rate}{}
  \item{Raw_cost_estimate_2017_USD_PPP}{}
  \item{Cost_estimate_per_year_local_currency}{}
  \item{Cost_estimate_per_year_2017_USD_exchange_rate}{}
  \item{Cost_estimate_per_year_2017_USD_PPP}{}
  \item{Currency}{}
  \item{Applicable_year}{}
  \item{Type_of_applicable_year}{}
  \item{Implementation}{}
  \item{Acquisition_method}{}
  \item{Impacted_sector}{}
  \item{Type_of_cost}{}
  \item{Method_reliability}{}
  \item{Details}{}
  \item{Contributors}{}
  \item{Probable_starting_year_low_margin}{}
  \item{Probable_ending_year_low_margin}{}
}}
\source{
Include URL to the original dataset here
}
\usage{
invacost
}
\description{
The database compiling published values of economic costs of Invasive Alien
Species.
}
\references{
Include both the data paper and the Nature paper here.
}
\keyword{datasets}
