% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfunctions.R
\name{plot.invacost.trendcost}
\alias{plot.invacost.trendcost}
\title{Plot model predictions of cost trends over time}
\usage{
\method{plot}{invacost.trendcost}(
  x,
  plot.breaks = 10^(-15:15),
  plot.type = "facets",
  graphical.parameters = NULL,
  ...
)
}
\arguments{
\item{x}{The output object from \code{\link{costTrendOverTime}}}

\item{plot.breaks}{a vector of numeric values indicating the plot breaks 
for the Y axis (cost values)}

\item{plot.type}{\code{"single"} or \code{"facets"}. Defines the type of plot
you want to make: a single facet with all models (\code{"single"}), or a 
facet per category of model (\code{"facets"})}

\item{graphical.parameters}{set this to \code{"manual"} if you want to 
customise ggplot2 parameters. 
By default, the following layers are configured: ylab, xlab, scale_x_continuous,
theme_bw and, if \code{cost.transf = "log10"}, scale_y_log10 and 
annotation_logticks. If you specify \code{grahical.parameters = "manual"},
all defaults will be ignored.}

\item{...}{additional arguments, none implemented for now}
}
\description{
This function provides different plotting methods for the estimated annual
cost of invasive species based on the temporal trend of costs.
}
\note{
If the legend appears empty (no colours) on your computer screen, try to
zoom in the plot, or to write to a file. There is a rare bug where under
certain conditions you cannot see the colours in the legend, because of their
transparency; zooming in or writing to a file are the best workarounds.
}
\examples{
data(invacost)
db.over.time <- expandYearlyCosts(invacost,
                                  startcolumn = "Probable_starting_year_low_margin",
                                  endcolumn = "Probable_ending_year_low_margin")
costdb <- db.over.time[db.over.time$Implementation == "Observed", ]
costdb <- costdb[which(costdb$Method_reliability == "High"), ]
res <- costTrendOverTime(costdb)
plot(res)
plot(res, plot.type = "single")
}
