% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRawCosts.R
\name{rawAvgCost}
\alias{rawAvgCost}
\title{Calculate the raw average annual cost over a single period of time}
\usage{
rawAvgCost(
  costdb,
  cost.column = "Cost_estimate_per_year_2017_USD_exchange_rate",
  year.column = "Impact_year",
  min.year = NULL,
  max.year = NULL
)
}
\arguments{
\item{costdb}{The \bold{expanded INVACOST database} output from 
\code{\link{expandYearlyCosts}},
where annual costs occurring over several years are repeated for each year.}

\item{cost.column}{Name of the cost column to use in \code{costdb} (usually, 
choose between the exchange rate (default) or PPP annualised cost)}

\item{year.column}{Name of the year column to use in \code{costdb}.}

\item{min.year}{The minimum year of the period (specify only if different from
the range of data)}

\item{max.year}{The minimum year of the period (specify only if different from
the range of data)}
}
\value{
a named \code{list} with 5 elements
}
\description{
This simple function calculates the raw average annual cost of invasive species
over a single period of time
}
\details{

}
\examples{
data(invacost)
db.over.time <- expandYearlyCosts(invacost,
                                  startcolumn = "Probable_starting_year_low_margin",
                                  endcolumn = "Probable_ending_year_low_margin")
costdb <- db.over.time[db.over.time$Implementation == "Observed", ]
costdb <- costdb[which(costdb$Method_reliability == "High"), ]
res <- rawAvgCost(costdb)
res
}
\seealso{
\code{\link{expandYearlyCosts}} to get the database in appropriate format.
}
\author{
Boris Leroy \email{leroy.boris@gmail.com}

with help from C. Diagne & A.-C. Vaissière
}
