% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandYearlyCosts.R
\name{expandYearlyCosts}
\alias{expandYearlyCosts}
\title{Expand annual cost values of the database over periods}
\usage{
expandYearlyCosts(costdb, startcolumn, endcolumn)
}
\arguments{
\item{costdb}{The \bold{INVACOST database}, usually obtained with
\code{data(\link{invacost})}.}

\item{startcolumn}{Name of the column containing starting years.}

\item{endcolumn}{Name of the column containing ending years.}
}
\value{
a \code{data.frame} containing the INVACOST database where 
all costs occuring over several years will be repeated for each year.
}
\description{
For costs that occur over several years, this function will repeat the 
annualised cost value over each year of the relevant period.
}
\details{

}
\examples{
# Create an example stack with two environmental variables
data(invacost)
invacost <- invacost[-which(is.na(invacost$Annualised.cost.estimate..2017.USD.exchange.rate.)), ]
db.over.time <- expandYearlyCosts(invacost,
                                  startcolumn = "Probable.Starting.year.Low.margin",
                                  endcolumn = "Probable.Ending.year.Low.margin")
}
\author{
Boris Leroy \email{leroy.boris@gmail.com}

with help from C. Diagne & A.-C. Vaissière
}
