% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettySummary.R
\name{prettySummary}
\alias{prettySummary}
\title{Provide a pretty summary of model results}
\usage{
prettySummary(x)
}
\arguments{
\item{x}{output object from \code{\link{costTrendOverTime}}}
}
\description{
This function is useful for presenting the main results (coefficients, tests,
etc.) of models in a paper
}
\examples{
data(invacost)
db.over.time <- expandYearlyCosts(invacost,
                                  startcolumn = "Probable_starting_year_low_margin",
                                  endcolumn = "Probable_ending_year_low_margin")
costdb <- db.over.time[db.over.time$Implementation == "Observed", ]
costdb <- costdb[which(costdb$Method_reliability == "High"), ]
costdb <- costdb[-which(is.na(costdb$Cost_estimate_per_year_2017_USD_exchange_rate)), ]
res <- costTrendOverTime(costdb)
prettySummary(res)
}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
